// src/modules/transfer/models/response/TransferResponse.ts
import {FulfillmentResponse} from './FulfillmentResponse';

export interface LineItemResponse {
  id: number;
  code: string;
  version: number;
  createdBy: string;
  createdName: string;
  createdDate: string;
  updatedBy: string;
  updatedName: string;
  updatedDate: string;
  sku: string;
  barcode: string;
  variantName: string;
  variantId: number;
  variantImage: string | null;
  productName: string;
  productId: number;
  available: number;
  transferQuantity: number;
  realQuantity: number;
  amount: number;
  price: number;
  weight: number;
  weightUnit: string;
  receiveOnHand: number;
}

export interface TransferResponse {
  id: number;
  code: string;
  version: number;
  createdBy: string;
  createdName: string;
  createdDate: Date;
  updatedBy: string;
  updatedName: string;
  updatedDate: string;
  fromStoreId: number;
  fromStoreCode: string;
  fromStoreName: string;
  fromStorePhone: string;
  fromStoreAddress: string;
  toStoreId: number;
  toStoreCode: string;
  toStoreName: string;
  toStorePhone: string;
  toStoreAddress: string;
  status: string;
  totalSentVariant: number;
  totalSentQuantity: number;
  totalSentAmount: number;
  requestedDate: string | null;
  requestedBy: string;
  requestedName: string;
  confirmedDate: string;
  confirmedBy: string;
  confirmedName: string;
  transferDate: string;
  transferBy: string;
  transferName: string;
  arrivedDate: string;
  arrivedBy: string;
  arrivedName: string;
  receiveDate: string;
  receivedBy: string;
  receivedName: string;
  pendingDate: string | null;
  pendingBy: string | null;
  pendingName: string | null;
  cancelDate: string | null;
  cancelBy: string | null;
  cancelName: string | null;
  balanceDate: string | null;
  balanceBy: string | null;
  balanceName: string | null;
  attachedFiles: string[];
  note: string | null;
  forwardNote: string | null;
  totalReceivedQuantity: number;
  lineItems: LineItemResponse[];
  shipment: any | null;
  totalReceiveOnHand: number | null;
  receivedMethod: string;
  subStatus: any | null;
  fulfillments: FulfillmentResponse[];
}